Add my initial documented disassembly of Faxanadu for the NES.
I have been working on disassembling and documenting one of my favorite
NES games, Faxanadu. The goal is to have a full understanding of how the
game works, the purpose of every used byte of memory, and to uncover any
secrets still left in the game.

I've been largely successful here, but there's still much work to do.

This initial disassembly has all code blocks identified, names and
documentation for the key areas of the game.

There are 4 deliverables in this change:

1. Ghidra files where the main disassembly work is being done.
2. Plain text assembly files for the game, using a custom exporter.
3. HTML-formatted assembly
4. Mesen symbols files.

The assembly file can be recompiled by asm6f, yielding the original
Faxanadu ROM.