Document the wired-off Hit Wall effect for Deluge and Fire spells.
Turns out that the Deluge and Fire spells were meant to have an effect
when the spell hit a wall, resulting in a brief explosion. This is a
really nice effect, though at least for the Fire spell it sometimes
doesn't show up. They may have decided it wasn't worth fixing, but left
all the code in, and just wired up the spells to disappear instead.

These can be reactivated with the following Game Genie codes:

ZAVLOUNN
ZEUUSUNN