Switch compilation to ca65.
The cc65 suite, and ca65 assembler, provide wider platform and language
feature support, making it more feasible to begin porting Faxanadu to
other platforms or to better create modifications based off the source.

The generated and annotated assembly will now be in ca65 format going
forward.

Along with this migration, the generator has better support for
converting addresses in instructions to references, as part of an effort
to produce a more relocatable ROM.